refactor(test): cleanup test utilities
