+++ tests/specs/plugins_load_spec.lua
